+++ source/spritetools_input.c
@@ -13,7 +13,7 @@
 static touchPosition INPUTTouchOrigin;
 static touchPosition INPUTTouchPosition;
 static touchPosition INPUTTouchLastPosition;
-static unsigned int INPUTTouchLength;
+static u32 INPUTTouchLength;
 static circlePosition INPUTCirclePosition;
 static accelVector INPUTAccelVector;
 static angularRate INPUTGyroRate;
@@ -23,7 +23,7 @@ static u32 INPUTKeysUp;
 
 /* Enables the Accelerometer and Gyroscope at startup */
 /* Returns 1 on success and 0 on failure */
-int ST_InputInit(void)
+u8 ST_InputInit(void)
 {
   HIDUSER_EnableAccelerometer();
   HIDUSER_EnableGyroscope();
@@ -68,7 +68,7 @@ void ST_InputScan(void)
 /* Checks for if a button was just pressed. Requires ST_InputScan before it */
 /* Takes key value enum from ctrulib to check */
 /* Returns 1 if the button has just been pressed. 0 if not */
-int ST_InputButtonPressed(u32 key)
+u8 ST_InputButtonPressed(u32 key)
 {
   if (INPUTKeysDown & key)
     return 1;
@@ -78,7 +78,7 @@ int ST_InputButtonPressed(u32 key)
 /* Checks for if a button is being pressed. Requires ST_InputScan before it */
 /* Takes key value enum from ctrulib to check */
 /* Returns 1 if the button is currently down. 0 if not */
-int ST_InputButtonDown(u32 key)
+u8 ST_InputButtonDown(u32 key)
 {
   if (INPUTKeysHeld & key)
     return 1;
@@ -88,7 +88,7 @@ int ST_InputButtonDown(u32 key)
 /* Checks for if a button was just released. Requires ST_InputScan before it */
 /* Takes key value enum from ctrulib to check */
 /* Returns 1 if the button was just released. 0 if not */
-int ST_InputButtonReleased(u32 key)
+u8 ST_InputButtonReleased(u32 key)
 {
   if (INPUTKeysUp & key)
     return 1;
@@ -97,7 +97,7 @@ int ST_InputButtonReleased(u32 key)
 
 /* Returns X position of touch. Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchX(void)
+s16 ST_InputTouchX(void)
 {
   if (INPUTKeysHeld & KEY_TOUCH)
     return INPUTTouchPosition.px;
@@ -106,7 +106,7 @@ int ST_InputTouchX(void)
 
 /* Returns Y position of touch. Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchY(void)
+s16 ST_InputTouchY(void)
 {
   if (INPUTKeysHeld & KEY_TOUCH)
     return INPUTTouchPosition.py;
@@ -116,7 +116,7 @@ int ST_InputTouchY(void)
 /* Returns X position of where the touch started. */
 /*   Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchOriginX(void)
+s16 ST_InputTouchOriginX(void)
 {
   if (INPUTKeysHeld & KEY_TOUCH)
     return INPUTTouchOrigin.px;
@@ -126,7 +126,7 @@ int ST_InputTouchOriginX(void)
 /* Returns Y position of where the touch started. */
 /*   Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchOriginY(void)
+s16 ST_InputTouchOriginY(void)
 {
   if (INPUTKeysHeld & KEY_TOUCH)
     return INPUTTouchOrigin.py;
@@ -136,7 +136,7 @@ int ST_InputTouchOriginY(void)
 /* Returns X difference between where the touch is and where it started. */
 /*   Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchDistanceX(void)
+s16 ST_InputTouchDistanceX(void)
 {
   if (INPUTKeysHeld & KEY_TOUCH)
     return abs(INPUTTouchPosition.px - INPUTTouchOrigin.px);
@@ -147,7 +147,7 @@ int ST_InputTouchDistanceX(void)
 /* Returns Y difference between where the touch is and where it started. */
 /*   Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchDistanceY(void)
+s16 ST_InputTouchDistanceY(void)
 {
   if (INPUTKeysHeld & KEY_TOUCH)
     return abs(INPUTTouchPosition.py - INPUTTouchOrigin.py);
@@ -158,7 +158,7 @@ int ST_InputTouchDistanceY(void)
 /* Returns distance between where the touch is and where it started. */
 /*   Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchDistance(void)
+s16 ST_InputTouchDistance(void)
 {
   if (INPUTKeysHeld & KEY_TOUCH)
     return sqrt(ST_InputTouchDistanceX() * ST_InputTouchDistanceX() +
@@ -170,7 +170,7 @@ int ST_InputTouchDistance(void)
 /* Returns length of line drawn on touchscreen. */
 /*   Requires ST_InputScan before it */
 /* Returns -1 if the touch screen is currently not being touched */
-int ST_InputTouchLength(void)
+s64 ST_InputTouchLength(void)
 {
   return INPUTTouchLength;
 }
